# nursery-inventory-system/nursery-inventory-system/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /inventory-items:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List current inventory items with stock levels and computed costs
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: plantId
          schema: { type: integer }
        - in: query
          name: nurseryLocation
          schema: { type: string }
        - in: query
          name: sortBy
          schema: { type: string, enum: [stock_asc, stock_desc] }
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory-items/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item with batch breakdown
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Inventory item detail with batch breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryDetail"
  /inventory-items/{id}/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      summary: Paginated inventory movement ledger for one item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: dateFrom
          schema: { type: string, format: date }
        - in: query
          name: dateTo
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated list of inventory movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInventoryMovements"
  /inventory-items/{id}/movements/export:
    get:
      operationId: exportInventoryMovements
      tags: [inventory]
      summary: Export movement ledger as CSV
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: dateFrom
          schema: { type: string, format: date }
        - in: query
          name: dateTo
          schema: { type: string, format: date }
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema: { type: string }
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Plant Master schemas ─────────────────────────────────────────────────
...
    InventoryItemPlant:
      type: object
      required: [id, plantCode, commonName, botanicalName, potSize]
      properties:
        id:            { type: integer }
        plantCode:     { type: string }
        commonName:    { type: string }
        botanicalName: { type: string }
        potSize:       { type: string }
        imageUrl:      { type: string, nullable: true }
    InventoryItem:
      type: object
      required:
        - id
        - plantId
        - plant
        - potSize
        - nurseryLocation
        - currentStock
        - costPerPlant
        - inventoryValue
        - updatedAt
      properties:
        id:              { type: integer }
        plantId:         { type: integer }
        plant:
          $ref: "#/components/schemas/InventoryItemPlant"
        potSize:         { type: string }
        nurseryLocation: { type: string }
        currentStock:    { type: integer }
        costPerPlant:    { type: number }
        inventoryValue:  { type: number }
        updatedAt:       { type: string, format: date-time }
    InventoryPurchaseBatch:
      type: object
      required:
        - batchNumber
        - supplier
        - purchaseDate
        - quantityPurchased
        - currentQuantity
        - issuedQuantity
        - costPerPlant
        - totalBatchCost
      properties:
        batchNumber:       { type: string }
        supplier:          { type: string }
        purchaseDate:      { type: string, format: date }
        quantityPurchased: { type: integer }
        currentQuantity:   { type: integer }
        issuedQuantity:    { type: integer }
        costPerPlant:      { type: number }
        totalBatchCost:    { type: number }
    InventoryProductionBatch:
      type: object
      required:
        - batchNumber
        - productionType
        - productionDate
        - successfulPlants
        - currentQuantity
        - issuedQuantity
        - costPerPlant
        - totalProductionCost
      properties:
        batchNumber:         { type: string }
        productionType:      { type: string }
        productionDate:      { type: string, format: date }
        successfulPlants:    { type: integer }
        currentQuantity:     { type: integer }
        issuedQuantity:      { type: integer }
        costPerPlant:        { type: number }
        totalProductionCost: { type: number }
    InventoryDetail:
      type: object
      required:
        - id
        - plantId
        - plant
        - potSize
        - nurseryLocation
        - currentStock
        - costPerPlant
        - inventoryValue
        - updatedAt
        - purchaseBatches
        - productionBatches
      properties:
        id:              { type: integer }
        plantId:         { type: integer }
        plant:
          $ref: "#/components/schemas/InventoryItemPlant"
        potSize:         { type: string }
        nurseryLocation: { type: string }
        currentStock:    { type: integer }
        costPerPlant:    { type: number }
        inventoryValue:  { type: number }
        updatedAt:       { type: string, format: date-time }
        purchaseBatches:
          type: array
          items:
            $ref: "#/components/schemas/InventoryPurchaseBatch"
        productionBatches:
          type: array
          items:
            $ref: "#/components/schemas/InventoryProductionBatch"
    InventoryMovementIssue:
      type: object
      required: [issueNumber, projectCode, projectName]
      properties:
        issueNumber: { type: string }
        projectCode: { type: string }
        projectName: { type: string }
    InventoryMovementRecord:
      type: object
      required:
        - id
        - movementType
        - batchNumber
        - batchSource
        - quantity
        - costPerPlant
        - totalCost
        - movementDate
      properties:
        id:           { type: integer }
        movementType: { type: string }
        batchNumber:  { type: string }
        batchSource:  { type: string }
        quantity:     { type: integer }
        costPerPlant: { type: number }
        totalCost:    { type: number }
        movementDate: { type: string, format: date }
        plantIssueId: { type: integer, nullable: true }
        plantIssue:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/InventoryMovementIssue"
    PaginatedInventoryMovements:
      type: object
      required: [data, total, page, pageSize, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryMovementRecord"
        total:      { type: integer }
        page:       { type: integer }
        pageSize:   { type: integer }
        totalPages: { type: integer }
...
    UserRole:
      type: string
      enum:
        - Administrator
        - Manager
        - Accountant
        - InventoryController
        - NurseryStaff
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, minLength: 1 }
    AuthUser:
      type: object
      required: [id, email, name, role, isActive, createdAt, updatedAt]
      properties:
        id:        { type: integer }
        email:     { type: string }
        name:      { type: string }
        role:      { $ref: "#/components/schemas/UserRole" }
        isActive:  { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...